+++ documentation/src/site/sphinx/release_notes.rst
@@ -7,6 +7,7 @@ Release Notes
 * Feature     : Contracts can update the filters on the server
 * Feature     : ContractServer can now be stopped via function call
 * Improvement : Git checkout code will now reuse existing checkouts and not clone every time
+* Bug fix     : Pattern matching for urls matched all of a url not ignoring '/'
 
 0.0.7
 -----